+++ pages/api/admin/committee/delete.ts
@@ -1,9 +1,17 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { PrismaClient } from "@prisma/client";
+import * as jose from "jose";
 
 const prisma = new PrismaClient();
 
 export default async function add(req: NextApiRequest, res: NextApiResponse) {
+
+  const {payload} = await jose.jwtVerify(req.cookies.token || "", new TextEncoder().encode(process.env.PASSWORD))
+
+  if (!payload) {
+    res.status(401).json({ message: "Unauthorized" })
+    return
+  }
 
   const year = req.body.year
